// src/types/searchProduct.type.ts
/**
 * Author: Nguyen Dinh Lam
 * Email: kiminonawa1305@gmail.com
 * Phone number: +84 855354919
 * Create at: 6:53 PM - 18/11/2024
 * User: lam-nguyen
 **/

type SearchProductType = {
	name?: string;
	category?: String;
	isBestSeller?: boolean;
	isNew?: boolean;
	page?: number;
};

export default SearchProductType;